+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **(app)** CD+Graphics backgrounds are now content-aware for better visuals and readability
 - **(app)** Summary is now shown following a media scan
 - **(server)** Support for custom URL base path (see `--urlPath` CLI option)
+- **(server)** Reduced load and improved client connection speed via library caching
 
 ### Changed